+++ CHANGELOG.md
@@ -34,7 +34,12 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 ### Changed
 
+- Basic Authentication failure logs now include the client IP address
+  ([#8428]).
+
 - The `edge` channel has been switched to the new UI and versioning scheme.
+
+[#8428]: https://github.com/AdguardTeam/AdGuardHome/issues/8428
 
 ### Deprecated